% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobb_douglas.R
\name{cobb_douglas}
\alias{cobb_douglas}
\title{Cobb-Douglas Model}
\usage{
cobb_douglas(I, Elas, K = 1, log = FALSE)
}
\arguments{
\item{I}{is a vector of inputs}

\item{Elas}{is a vector of elasticities, must be the same length as I.}

\item{K}{is the constant of the model. Defaults to 1.}

\item{log}{is logical. If TRUE, then output is in natural logarithm. Defaults to FALSE.}
}
\description{
This function allows you to compute a Cobb-Douglas production/ utility function with n inputs/goods.
}
\details{
cobb_douglas_2 computes what - mathematically - is a particular case of this function, but computationally there are differentes. Here, the user must input two vectors, one for elasticies and one for quantities, whereas in cobb_douglas_2, the user specifies only quantities and elasticities are taken as parameters.
}
