% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost2.R
\name{cost2}
\alias{cost2}
\title{Cost function}
\usage{
cost2(q2 = "q2", q0_2 = 0, c2 = 1, c2_2 = 0)
}
\arguments{
\item{q2}{is output.}

\item{q0_2}{is the firm's fixed cost. Defatults to 0.}

\item{c2}{is linear marginal cost. Defatults to 1.}

\item{c2_2}{is the quadratic term's coefficient. Defaults to zero.}
}
\description{
This function provides a general cost curve in a symbolic of numeric manner. Needed to use some functions of this package.
}
\examples{
cost2()
}
\keyword{competition}
\keyword{imperfect}
\keyword{microeconomics}
