% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cournot_2.R
\name{cournot_2}
\alias{cournot_2}
\title{Cournot Duopoly}
\usage{
cournot_2(q0_1 = 0, c1 = 1, c1_2 = 0, q0_2 = 0, c2 = 1,
  c2_2 = 0, p_0, a = 1, b = 0)
}
\arguments{
\item{firm1}{is a cost function. cost1 is provided in this package.}

\item{firm2}{is a cost function. cost2 is provided in this package.}

\item{demand}{is an inverse demand curve. demand is also provided in this package.}
}
\description{
This function solves analitically a Cournot Duopoly model. One should provide demand and cost functions and the function returns the solution to the problem.
}
\examples{
cournot_2()
}
\keyword{competition}
\keyword{imperfect}
\keyword{microeconomics}
