% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{demand}
\alias{demand}
\title{Inverse demand curve}
\usage{
demand(q = Var("q"), p_0, a = 1, b = 0)
}
\arguments{
\item{q}{is total output.}

\item{p_0}{is the demand curve intercept.}

\item{a}{is dq/dP. Defaults to 1.}

\item{b}{is the response to the quadratic q term. Defaults to 0.}
}
\description{
This function provides a general inverse demand curve in a symbolic of numeric manner. Needed to use some functions of this package.
}
\examples{
demand()
}
\keyword{competition}
\keyword{imperfect}
\keyword{microeconomics}
