% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis2.R
\name{get_data2}
\alias{get_data2}
\title{Get data from downloaded XLSX from IDI2}
\usage{
get_data2(file, sheet = NULL, ...)
}
\arguments{
\item{file}{character, XLSX file with data}

\item{sheet}{numeric/character, index number of name of the sheet(s) to be loaded}

\item{...}{other arguments passed to \code{\link{readWorksheet}}}
}
\value{
Data frame or list thereof.
}
\description{
By default retrieves all sheets from interviews and returns a list of data frames.
}

