% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRW_steady_state.R
\name{MRW_steady_state}
\alias{MRW_steady_state}
\title{Mankiw-Romer-Weil Growth Model Steady State}
\usage{
MRW_steady_state(n = 0.01, g = 0.01, alpha = 0.33, beta = 0.33,
  sk = 0.01, sh = 0.01, delta = 0.01, gamma = 0)
}
\arguments{
\item{n}{is population growth rate. Defaults to 1%.}

\item{g}{is the technological growth rate. Defaults to 1%.}

\item{alpha}{is capital-output elasticity. Defaults to .33 as estimated by Mankiw, Romer and Weil.}

\item{beta}{is the human capital-output elasciticy. Defatults to .33 as estimated by Mankiw, Romer and Weil.}

\item{sk}{is the savings rate devoted to physical capital. Defaults to 1%.}

\item{sh}{is the savings rate devoted to human capital. Defaults to 1%.}

\item{delta}{is the physical capital stock's depreciation rate. Defaults to 1%.}

\item{gamma}{is the human capital stock's depreciation rate. Defaults to 0.}
}
\description{
This function computes steady state income, capital and human capital per worker given relevant parameters according to the modelo presented in Mankiw, Romer and Weil (1992).
}
\examples{
demand()
}
\keyword{growth}
\keyword{macroeconomics}
\keyword{theory}
