% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous_games_mixed.R
\name{sgms}
\alias{sgms}
\title{Simultaneous Games Mixed Strategies}
\usage{
sgms(a, b)
}
\arguments{
\item{a}{is the row player's payoff matrix.}

\item{b}{is the column player's payoff matrix.}
}
\description{
This function finds the Nash equilibrium in mixed strategies of a 2-person simultaneous game. The function works only for 2x2 matrices,
}
\keyword{microeconomic.}
