% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous_games_pure.R
\name{sim_pure_nasheq}
\alias{sim_pure_nasheq}
\title{Simultaneous Games Pure Strategies}
\usage{
sim_pure_nasheq(a, b)
}
\arguments{
\item{a}{is the row player's payoff matrix.}

\item{b}{is the column player's payoff matrix.}
}
\description{
This function finds the Nash equilibrium in pure strategies of a 2-person simultaneous game.
}
