% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{vis}
\title{Visualise net from sheet}
\usage{
vis(sheet)
}
\arguments{
\item{sheet}{sheet name}
}
\description{
Visualise net from sheet.
}
\details{
Needs authorisation with \code{gs_auth()} first.
}
\examples{
\dontrun{
library(googlesheets)
gs_auth()
vis("WF01")
}
}

