% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis2.R
\name{vis_net2}
\alias{vis_net2}
\title{Plot IDI2 network}
\usage{
vis_net2(g, show_groups = NULL, ...)
}
\arguments{
\item{g}{igraph object}

\item{show_groups}{numeric, ids of groups to display. Alternatively, NULL for all, NA for none,}

\item{...}{passed to plot.igraph}
}
\value{
nothing
}
\description{
Plot IDI2 network
}

