% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cournot_2.R
\name{cournot_2}
\alias{cournot_2}
\title{Cournot Duopoly}
\usage{
cournot_2(firm1, firm2, demand)
}
\arguments{
\item{firm1}{is a cost function. cost1 is provided in this package.}

\item{firm2}{is a cost function. cost2 is provided in this package.}

\item{demand}{is an inverse demand curve. demand is also provided in this package.}
}
\description{
This function a Cournot duopoly model. One should provide demand and cost functions and the function returns the solution to the problem.
}
\examples{
cournot_2()
}
\keyword{competition}
\keyword{imperfect}
\keyword{microeconomics}
