% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{make_net}
\alias{vis}
\alias{vis_all}
\alias{vis_net}
\title{Visualise team networks}
\usage{
vis(sheet, ...)

make_net(d)

vis_net(g, vid = V(g), gid = "all", vcol = c("#66c2a5", "yellow",
  "#8da0cb", "white"), vshape = c("mbcircle", "mbsquare"), vframe = c(black
  = NA, `#e41a1c` = 0, `#377eb8` = 1), gcol = RColorBrewer::brewer.pal(8,
  "Set3"), ggroups = TRUE)

vis_all(sheet_names = NULL, skip_sheets = NULL, overwrite = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{sheet}{sheet name}

\item{...}{other arguments passed to \code{vis_net}}

\item{d}{data frame downloaded from GS}

\item{g}{igraph object}

\item{vid}{numeric ids of vertices to draw, all by default}

\item{gid}{numeric ids of groups to draw or \code{"all"} (default), or \code{"none"}.}

\item{vcol}{vertex color mapping}

\item{vshape}{vertex shape mapping}

\item{vframe}{vertex frame color mapping}

\item{gcol}{vertex group color mapping}

\item{ggroups}{whether to take groups into account when calculating layout}

\item{sheet_names}{names of GS sheet names to visualize, if \code{NULL} draws all \code{WF} and \code{DC} sheets with numbers 1 to 15. Single character element (e.g. \code{"WF"}) is interpreted as a regular expression to search through sheet names}

\item{skip_sheets}{character vector, if \code{sheet_names} is \code{NULL}, sheets with these names will be skipped}

\item{overwrite}{logical redo existing files}

\item{verbose}{logical give verbose output or not}
}
\description{
Download and visualize team networks based on data in a Google Spreadsheet.

\code{make_net} downloads data and creates igraph object.

\code{vis_net} draws created igraph object.

\code{vis_all} visualizes all networks and saves them to files 'XXXX.png'
}
\details{
Needs authorisation with \code{gs_auth()} first.
}

